fix np.float issue with newer numpy versions (newer than current supply by Blender)

+++ utils/mesh.py
@@ -10,7 +10,7 @@ def get_bbox(mesh=None, coords=None):
     '''
 
     vert_count = len(mesh.vertices)
-    coords = np.empty((vert_count, 3), np.float)
+    coords = np.empty((vert_count, 3), float)
     mesh.vertices.foreach_get('co', np.reshape(coords, vert_count * 3))
 
     # find extreme values on each axis in both directions
@@ -57,19 +57,19 @@ def get_coords(mesh, mx=None, offset=0, indices=False):
     verts = mesh.vertices
     vert_count = len(verts)
 
-    coords = np.empty((vert_count, 3), np.float)
+    coords = np.empty((vert_count, 3), float)
     mesh.vertices.foreach_get('co', np.reshape(coords, vert_count * 3))
 
     # offset along vertex normal
     if offset:
-        normals = np.empty((vert_count, 3), np.float)
+        normals = np.empty((vert_count, 3), float)
         mesh.vertices.foreach_get('normal', np.reshape(normals, vert_count * 3))
 
         coords = coords + normals * offset
 
     # bring coords into non-local space
     if mx:
-        coords_4d = np.ones((vert_count, 4), dtype=np.float)
+        coords_4d = np.ones((vert_count, 4), dtype=float)
         coords_4d[:, :-1] = coords
 
         coords = np.einsum('ij,aj->ai', mx, coords_4d)[:, :-1]
